requirements.txt
authlib==1.6.9
deep_translator==1.11.4
fitz==0.0.1.dev2
Flask==3.1.3
Flask_Login==0.6.3
flask_mail==0.10.0
Flask_Migrate==4.1.0
flask_sqlalchemy==3.1.1
flask_wtf==1.2.2
googletrans==4.0.2
numpy==2.4.2
opencv_python_headless==4.13.0.92
openpyxl==3.1.5
pandas==3.0.1
pdfplumber==0.11.9
Pillow==12.1.1
pypdf==6.7.5
pytesseract==0.3.13
python-dotenv==1.2.2
python_docx==1.2.0
python_pptx==1.0.2
pytz==2026.1.post1
reportlab==4.4.10
SQLAlchemy==2.0.48
Werkzeug==3.1.6
